+++ crates/wdl-lint/src/rules/misplaced_lint_directive.rs
@@ -20,6 +20,7 @@ use wdl_ast::Visitor;
 use crate::Rule;
 use crate::Tag;
 use crate::TagSet;
+use crate::optional_rules;
 use crate::rules;
 
 /// The identifier for the unknown rule rule.
@@ -58,6 +59,10 @@ pub static RULE_MAP: LazyLock<HashMap<&'static str, Option<&'static [SyntaxKind]
     LazyLock::new(|| {
         let mut map = HashMap::new();
         for rule in rules() {
+            map.insert(rule.id(), rule.exceptable_nodes());
+        }
+        // insert optional rules as well
+        for rule in optional_rules() {
             map.insert(rule.id(), rule.exceptable_nodes());
         }
         map