+++ crates/wdl/src/bin/wdl.rs
@@ -51,6 +51,7 @@ use wdl_engine::local::LocalTaskExecutionBackend;
 use wdl_engine::v1::TaskEvaluator;
 use wdl_format::Formatter;
 use wdl_format::element::node::AstNodeFormatExt as _;
+use wdl_lint::rules::ShellCheckRule;
 
 /// Emits the given diagnostics to the output stream.
 ///
@@ -296,6 +297,9 @@ pub struct LintCommand {
     /// The path to the source WDL file.
     #[clap(value_name = "PATH")]
     pub path: PathBuf,
+    /// Enable shellcheck lints.
+    #[clap(long, action)]
+    pub shellcheck: bool,
 }
 
 impl LintCommand {
@@ -315,6 +319,9 @@ impl LintCommand {
 
         let mut validator = Validator::default();
         validator.add_visitor(LintVisitor::default());
+        if self.shellcheck {
+            validator.add_visitor(ShellCheckRule);
+        }
         if let Err(diagnostics) = validator.validate(&document) {
             emit_diagnostics(&self.path.to_string_lossy(), &source, &diagnostics)?;